Working windows however has the utf8 bug will fix later
