--- pip_show.log
@@ -1,20 +1,0 @@
-Name: transformers
-Version: 4.46.3
-Summary: State-of-the-art Machine Learning for JAX, PyTorch and TensorFlow
-Home-page: https://github.com/huggingface/transformers
-Author: The Hugging Face team (past and future) with the help of all our contributors (https://github.com/huggingface/transformers/graphs/contributors)
-Author-email: transformers@huggingface.co
-License: Apache 2.0 License
-Location: C:\Users\matth\Desktop\EpubTranslator\.venv\Lib\site-packages
-Requires: filelock, huggingface-hub, numpy, packaging, pyyaml, regex, requests, safetensors, tokenizers, tqdm
-Required-by: optimum
-Name: transformers
-Version: 4.46.3
-Summary: State-of-the-art Machine Learning for JAX, PyTorch and TensorFlow
-Home-page: https://github.com/huggingface/transformers
-Author: The Hugging Face team (past and future) with the help of all our contributors (https://github.com/huggingface/transformers/graphs/contributors)
-Author-email: transformers@huggingface.co
-License: Apache 2.0 License
-Location: C:\Users\matth\Desktop\EpubTranslator\.venv\Lib\site-packages
-Requires: filelock, huggingface-hub, numpy, packaging, pyyaml, regex, requests, safetensors, tokenizers, tqdm
-Required-by: optimum
