--- translatedTags.txt
@@ -1,50 +1,0 @@
-7,0,Blessed is this wonderful world!
-7,1,Oh, my goddess.
-7,2,♪ ♪ ♪ The ♪ ♪ The ♪ ♪ the ♪ ♪ the ♪ the ♪ ♪ the ♪ the ♪ ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪ the ♪
-7,5,Sneaker Library, Han River.
-8,0,We forbid all or parts of this work to be duplicated, reprinted, distributed, or sent to the home page without permission. Also, we forbid the content of the work to be changed or tampered with without permission.
-8,1,You cannot hand over this work to a third party, either free of charge or free of charge, thanks to the regulations you accepted when you bought it.
-8,2,Image images, such as thumbnails showing the work, can be changed without notice when re-downing.
-8,4,The work is arranged by vertical writing.
-8,5,Also, the reading system you see can sometimes show a difference in display.
-9,0,COUNTS
-9,3,Prologue
-9,4,Chapter One, I want you to meet this self-named goddess and a new world.
-9,5,Chapter Two, I want you to give me a lot of treasure in my right hand!
-9,6,Chapter three, give this lake the best squeeze of its own name!
-9,7,Chapter four, you must end this fragile battle!
-9,8,Epilogue
-9,10,I'd like to thank Dr. Aquaa for buying a volume of ebooks.
-12,3,Welcome to the afterlife, Mr. Fujisawa. You've just died an unhappy life, but your life is over.
-12,4,In the white room, I was suddenly told such a thing.
-12,6,I don't know what's out of the ordinary.
-12,7,There was a small office desk and a small desk in the room, and the person who told me the end of my life was sitting in that chair.
-12,9,If there is a god-woman, she must be talking about the other person before her eyes.
-12,10,It's completely different from the pretty looks of the idol that you see on TV. It's very humanly distinct.
-12,11,A long, clear, water-colored hair that gives a soft impression.
-12,12,I wonder if I'm as old as I am.
-12,13,It doesn't come out too much, it's not too perfect, it's pale, it's pale, it's purple, it's white, it's white, it's plain, it's wrapped in a loose garment called a feather.
-12,14,The pretty girl stared at me with her hair's color and clear, water-colored eyes glancing her eyes, and her cheery cheers filled her eyes.
-12,17,...I remember the memory I had so far.
-12,19,.
-12,21,I usually didn't go to school, but today I went out unusually.
-12,22,I got up a strange early and lined up in order to get a popular video game and its first limited version of today's game.
-12,23,I hear the world is calling out that people like me are out of control.
-12,24,I got the game safely caught, and then I went home, and then I was going to go home to play the game, and then I was going to go home to get my hands on it.
-12,25,The girl who was walking right in front of me with her arms tied.
-12,26,From the way I dress, I wonder if I'm a student at the same school as I am.
-12,27,When the traffic light turned blue, the child went right across the street without looking right or left at Lok.
-12,28,It's a big shadow that's on the way to being next to a girl.
-12,29,It must have been a big truck moving fast.
-12,30,I pushed him before I thought about it.
-12,32,And...
-12,34,...with a strangely calm mind, I asked the pretty girl in front of me quietly.
-12,35,Can I ask you just one question?
-12,36,A pretty girl nodded at my question.
-12,37,Would you like that?
-12,38,The girl I pushed, is she alive?
-12,39,It was important.
-12,40,It was my first and last show in my life.
-12,41,I'm too sorry if I didn't get there in time.
-12,42,"He's alive, isn't he?"
-12,43,I'm glad...
